Sales Dashboard
This is my first power bi project which i was created by own. I was collected the data from online and it was a sample data. I had used all the  features which are provided by MS and i will my skills to get more efficiency while creating the dashboard.